requirements.txt
streamlit
pandas
openai
python-dotenv
loguru
langchain
langchain-openai
pydantic
requests